fix: User added tokens logos is not visible on token search modal list
