+++ src/components/SearchModal/CurrencySearch.tsx
@@ -7,7 +7,8 @@ import { FixedSizeList } from 'react-window'
 import { useAudioModeManager } from 'state/user/hooks'
 import useDebounce from 'hooks/useDebounce'
 import useActiveWeb3React from 'hooks/useActiveWeb3React'
-import { TagInfo, useAllLists, useInactiveListUrls, WrappedTokenInfo } from 'state/lists/hooks'
+import { useAllLists, useInactiveListUrls } from 'state/lists/hooks'
+import { TagInfo, WrappedTokenInfo } from 'state/types'
 import { useAllTokens, useToken, useIsUserAddedToken } from '../../hooks/Tokens'
 import { isAddress } from '../../utils'
 import Column, { AutoColumn } from '../Layout/Column'
