+++ src/state/user/hooks/helpers.ts
@@ -1,6 +1,6 @@
 import { Token } from '@pancakeswap/sdk'
 import { SerializedToken } from 'config/constants/types'
-import { parseUnits } from '@ethersproject/units'
+import { WrappedTokenInfo } from 'state/types'
 
 export function serializeToken(token: Token): SerializedToken {
   return {
@@ -10,10 +10,24 @@ export function serializeToken(token: Token): SerializedToken {
     symbol: token.symbol,
     name: token.name,
     projectLink: token.projectLink,
+    logoURI: token instanceof WrappedTokenInfo ? token.logoURI : undefined,
   }
 }
 
 export function deserializeToken(serializedToken: SerializedToken): Token {
+  if (serializedToken.logoURI) {
+    return new WrappedTokenInfo(
+      {
+        chainId: serializedToken.chainId,
+        address: serializedToken.address,
+        decimals: serializedToken.decimals,
+        symbol: serializedToken.symbol,
+        name: serializedToken.name,
+        logoURI: serializedToken.logoURI,
+      },
+      [],
+    )
+  }
   return new Token(
     serializedToken.chainId,
     serializedToken.address,
@@ -22,18 +36,4 @@ export function deserializeToken(serializedToken: SerializedToken): Token {
     serializedToken.name,
     serializedToken.projectLink,
   )
-}
-
-export enum GAS_PRICE {
-  default = '5',
-  fast = '6',
-  instant = '7',
-  testnet = '10',
-}
-
-export const GAS_PRICE_GWEI = {
-  default: parseUnits(GAS_PRICE.default, 'gwei').toString(),
-  fast: parseUnits(GAS_PRICE.fast, 'gwei').toString(),
-  instant: parseUnits(GAS_PRICE.instant, 'gwei').toString(),
-  testnet: parseUnits(GAS_PRICE.testnet, 'gwei').toString(),
 }
