+++ src/utils/getGasPrice.ts
@@ -1,7 +1,7 @@
 import { ChainId } from '@pancakeswap/sdk'
 import { CHAIN_ID } from 'config/constants/networks'
 import store from 'state'
-import { GAS_PRICE_GWEI } from 'state/user/hooks/helpers'
+import { GAS_PRICE_GWEI } from 'state/types'
 
 /**
  * Function to return gasPrice outwith a react component
